chore(*) rename/delete files
